requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.2
django-allauth==65.1.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
gunicorn==23.0.0
packaging==24.1
pillow==11.0.0
psycopg==3.1.12
sqlparse==0.5.1
typing_extensions==4.12.2
whitenoise==6.8.2
